Dodanie filtrowania

+++ SQLProjektV2/Views/GodzinyPracyView.xaml.cs
@@ -45,9 +45,15 @@ namespace SQLProjektV2.Views
             if (x != null)
             {
                 selectedId = x.Text;
-                OptionChoose.Visibility = Visibility.Visible;
                 AddForm.Visibility = Visibility.Collapsed;
-                ModForm.Visibility = Visibility.Collapsed;
+                ModForm.Visibility = Visibility.Visible;
+                Filters.Visibility = Visibility.Collapsed;
+
+                DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdGodziny Pracy]", int.Parse(selectedId));
+
+                MPSource.SelectedValue = temp.Rows[0][0];
+                MDatePicker1.SelectedDate = ((DateTime)temp.Rows[0][1]);
+                MGodzinySource.Text = temp.Rows[0][2].ToString();
             }
         }
 
@@ -63,24 +69,18 @@ namespace SQLProjektV2.Views
 
         private void AddFormVisible(object sender, RoutedEventArgs e)
         {
-            OptionChoose.Visibility = Visibility.Collapsed;
             AddForm.Visibility = Visibility.Visible;
             ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Collapsed;
+
         }
 
-        private void ModFormVisible(object sender, RoutedEventArgs e)
+        private void Cancel(object sender, RoutedEventArgs e)
         {
             AddForm.Visibility = Visibility.Collapsed;
-            OptionChoose.Visibility = Visibility.Collapsed;
-            ModForm.Visibility = Visibility.Visible;
-
-            DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdGodziny Pracy]", int.Parse(selectedId));
-
-            MPSource.SelectedValue = temp.Rows[0][0];
-            MDatePicker1.SelectedDate = ((DateTime)temp.Rows[0][1]);
-            MGodzinySource.Text = temp.Rows[0][2].ToString();
+            ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Visible;
         }
-
 
 
         private void AddNewRecord(object sender, RoutedEventArgs e)
@@ -149,19 +149,50 @@ namespace SQLProjektV2.Views
 
         private void DeleteRecord(object sender, RoutedEventArgs e)
         {
-            string temp = $"DELETE FROM [dbo].[Godziny_pracy] WHERE Id = {selectedId}";
-            try
-            {
-                DBConnection.SQLCommand(temp);
-                MessageBox.Show("Usunięto wpis godzin pracy");
-                DataContext = new GodzinyPracyViewModel();
-                OptionChoose.Visibility = Visibility.Collapsed;
-            }
-            catch (Exception)
-            {
-                MessageBox.Show("Nie usunięto wpisu godzin pracy, ponieważ jest związany z innymi tablicami");
-            }
-        }
+            MessageBoxResult result = MessageBox.Show("Czy na pewno chcesz usunąć wpis?", "",  MessageBoxButton.YesNo);
         
+            if(result == MessageBoxResult.Yes)
+            {
+                string temp = $"DELETE FROM [dbo].[Godziny_pracy] WHERE Id = {selectedId}";
+                try
+                {
+                    DBConnection.SQLCommand(temp);
+                    MessageBox.Show("Usunięto wpis godzin pracy");
+                    DataContext = new GodzinyPracyViewModel();
+                    ModForm.Visibility = Visibility.Collapsed;
+                    Filters.Visibility = Visibility.Visible;
+                }
+                catch (Exception)
+                {
+                    MessageBox.Show("Nie usunięto wpisu godzin pracy, ponieważ jest związany z innymi tablicami");
+                }
+            }
+
+        }
+
+
+
+        private void AddFilter(object sender, RoutedEventArgs e)
+        {
+            StackPanel temp = Globals.StackPanel;
+
+            ComboBox tempCB = Globals.ComboBoxColumnChoose;
+            tempCB.ItemsSource = ((GodzinyPracyViewModel)this.DataContext).FilterInfo;
+            temp.Children.Add(tempCB);
+
+            temp.Children.Add(Globals.ComboBox);
+
+            temp.Children.Add(Globals.Button);
+
+            (temp.Children[0] as ComboBox).SelectedIndex = 0;
+
+            FiltersList.Children.Add(temp);
+        }
+
+        private void UseFilters(object sender, RoutedEventArgs e)
+        {
+            (DataContext as GodzinyPracyViewModel).MainTable.RowFilter = Globals.GetFilter(FiltersList);
+        }
+
     }
 }
