+++ SQLProjektV2/Views/KlienciView.xaml.cs
@@ -45,9 +45,16 @@ namespace SQLProjektV2.Views
             if (x != null)
             {
                 selectedId = x.Text;
-                OptionChoose.Visibility = Visibility.Visible;
                 AddForm.Visibility = Visibility.Collapsed;
-                ModForm.Visibility = Visibility.Collapsed;
+                ModForm.Visibility = Visibility.Visible;
+                Filters.Visibility = Visibility.Collapsed;
+
+                DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdKlienci]", int.Parse(selectedId));
+
+                MImięSource.Text = temp.Rows[0][0].ToString();
+                MNazwiskoSource.Text = temp.Rows[0][1].ToString();
+                MEmailSource.Text = temp.Rows[0][2].ToString();
+                MNumerSource.Text = temp.Rows[0][3].ToString();
             }
         }
 
@@ -61,25 +68,17 @@ namespace SQLProjektV2.Views
 
         private void AddFormVisible(object sender, RoutedEventArgs e)
         {
-            OptionChoose.Visibility = Visibility.Collapsed;
             AddForm.Visibility = Visibility.Visible;
             ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Collapsed;
         }
 
-        private void ModFormVisible(object sender, RoutedEventArgs e)
+        private void Cancel(object sender, RoutedEventArgs e)
         {
             AddForm.Visibility = Visibility.Collapsed;
-            OptionChoose.Visibility = Visibility.Collapsed;
-            ModForm.Visibility = Visibility.Visible;
-
-            DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdKlienci]", int.Parse(selectedId));
-
-            MImięSource.Text = temp.Rows[0][0].ToString();
-            MNazwiskoSource.Text = temp.Rows[0][1].ToString();
-            MEmailSource.Text = temp.Rows[0][2].ToString();
-            MNumerSource.Text = temp.Rows[0][3].ToString();
+            ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Visible;
         }
-
 
 
         private void AddNewRecord(object sender, RoutedEventArgs e)
@@ -143,19 +142,48 @@ namespace SQLProjektV2.Views
 
         private void DeleteRecord(object sender, RoutedEventArgs e)
         {
-            string temp = $"DELETE FROM [dbo].[Klienci] WHERE Id = {selectedId}";
-            try
+            MessageBoxResult result = MessageBox.Show("Czy na pewno chcesz usunąć wpis?", "", MessageBoxButton.YesNo);
+
+            if (result == MessageBoxResult.Yes)
             {
-                DBConnection.SQLCommand(temp);
-                MessageBox.Show("Usunięto klienta");
-                DataContext = new KlienciViewModel();
-                OptionChoose.Visibility = Visibility.Collapsed;
-            }
-            catch (Exception)
-            {
-                MessageBox.Show("Nie usunięto klienta, ponieważ jest związany z innymi tablicami");
+                string temp = $"DELETE FROM [dbo].[Klienci] WHERE Id = {selectedId}";
+                try
+                {
+                    DBConnection.SQLCommand(temp);
+                    MessageBox.Show("Usunięto klienta");
+                    DataContext = new KlienciViewModel();
+                    ModForm.Visibility = Visibility.Collapsed;
+                    Filters.Visibility = Visibility.Visible;
+                }
+                catch (Exception)
+                {
+                    MessageBox.Show("Nie usunięto klienta, ponieważ jest związany z innymi tablicami");
+                }
             }
 
+        }
+
+
+        private void AddFilter(object sender, RoutedEventArgs e)
+        {
+            StackPanel temp = Globals.StackPanel;
+
+            ComboBox tempCB = Globals.ComboBoxColumnChoose;
+            tempCB.ItemsSource = ((KlienciViewModel)this.DataContext).FilterInfo;
+            temp.Children.Add(tempCB);
+
+            temp.Children.Add(Globals.ComboBox);
+
+            temp.Children.Add(Globals.Button);
+
+            (temp.Children[0] as ComboBox).SelectedIndex = 0;
+
+            FiltersList.Children.Add(temp);
+        }
+
+        private void UseFilters(object sender, RoutedEventArgs e)
+        {
+            (DataContext as KlienciViewModel).MainTable.RowFilter = Globals.GetFilter(FiltersList);
         }
 
     }
