+++ SQLProjektV2/Views/MiejscaView.xaml.cs
@@ -44,9 +44,15 @@ namespace SQLProjektV2.Views
             if (x != null)
             {
                 selectedId = x.Text;
-                OptionChoose.Visibility = Visibility.Visible;
                 AddForm.Visibility = Visibility.Collapsed;
-                ModForm.Visibility = Visibility.Collapsed;
+                ModForm.Visibility = Visibility.Visible;
+                Filters.Visibility = Visibility.Collapsed;
+
+
+                DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdMiejsca]", int.Parse(selectedId));
+
+                MAdresSource.Text = temp.Rows[0][0].ToString();
+                MNumerSource.Text = temp.Rows[0][1].ToString();
             }
         }
 
@@ -60,21 +66,17 @@ namespace SQLProjektV2.Views
 
         private void AddFormVisible(object sender, RoutedEventArgs e)
         {
-            OptionChoose.Visibility = Visibility.Collapsed;
             AddForm.Visibility = Visibility.Visible;
             ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Collapsed;
+
         }
 
-        private void ModFormVisible(object sender, RoutedEventArgs e)
+        private void Cancel(object sender, RoutedEventArgs e)
         {
             AddForm.Visibility = Visibility.Collapsed;
-            OptionChoose.Visibility = Visibility.Collapsed;
-            ModForm.Visibility = Visibility.Visible;
-
-            DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdMiejsca]", int.Parse(selectedId));
-
-            MAdresSource.Text = temp.Rows[0][0].ToString();
-            MNumerSource.Text = temp.Rows[0][1].ToString();
+            ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Visible;
         }
 
 
@@ -171,19 +173,47 @@ namespace SQLProjektV2.Views
 
         private void DeleteRecord(object sender, RoutedEventArgs e)
         {
-            string temp = $"DELETE FROM [dbo].[Miejsca] WHERE Id = {selectedId}";
-            try
-            {
-                DBConnection.SQLCommand(temp);
-                MessageBox.Show("Usunięto miejsce");
-                DataContext = new MiejscaViewModel();
-                OptionChoose.Visibility = Visibility.Collapsed;
-            }
-            catch (Exception)
-            {
-                MessageBox.Show("Nie usunięto miejsca, ponieważ jest związane z innymi tablicami");
-            }
+            MessageBoxResult result = MessageBox.Show("Czy na pewno chcesz usunąć wpis?", "", MessageBoxButton.YesNo);
 
+            if (result == MessageBoxResult.Yes)
+            {
+                string temp = $"DELETE FROM [dbo].[Miejsca] WHERE Id = {selectedId}";
+                try
+                {
+                    DBConnection.SQLCommand(temp);
+                    MessageBox.Show("Usunięto miejsce");
+                    DataContext = new MiejscaViewModel();
+                    ModForm.Visibility = Visibility.Collapsed;
+                    Filters.Visibility = Visibility.Visible;
+                }
+                catch (Exception)
+                {
+                    MessageBox.Show("Nie usunięto miejsca, ponieważ jest związane z innymi tablicami");
+                }
+            }
+        }
+
+
+        private void AddFilter(object sender, RoutedEventArgs e)
+        {
+            StackPanel temp = Globals.StackPanel;
+
+            ComboBox tempCB = Globals.ComboBoxColumnChoose;
+            tempCB.ItemsSource = ((MiejscaViewModel)this.DataContext).FilterInfo;
+            temp.Children.Add(tempCB);
+
+            temp.Children.Add(Globals.ComboBox);
+
+            temp.Children.Add(Globals.Button);
+
+            (temp.Children[0] as ComboBox).SelectedIndex = 0;
+
+            FiltersList.Children.Add(temp);
+        }
+
+        private void UseFilters(object sender, RoutedEventArgs e)
+        {
+            (DataContext as MiejscaViewModel).MainTable.RowFilter = Globals.GetFilter(FiltersList);
         }
     }
 }
