+++ SQLProjektV2/Views/ProjektyView.xaml.cs
@@ -45,9 +45,16 @@ namespace SQLProjektV2.Views
             if (x != null)
             {
                 selectedId = x.Text;
-                OptionChoose.Visibility = Visibility.Visible;
                 AddForm.Visibility = Visibility.Collapsed;
-                ModForm.Visibility = Visibility.Collapsed;
+                Filters.Visibility = Visibility.Collapsed;
+                ModForm.Visibility = Visibility.Visible;
+
+                DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdProjekty]", int.Parse(selectedId));
+
+                MDatePicker1.SelectedDate = ((DateTime)temp.Rows[0][0]);
+                MOpisSource.Text = temp.Rows[0][1].ToString();
+                MKSource.SelectedValue = temp.Rows[0][2];
+                MZSource.SelectedValue = temp.Rows[0][3];
             }
         }
 
@@ -63,25 +70,17 @@ namespace SQLProjektV2.Views
 
         private void AddFormVisible(object sender, RoutedEventArgs e)
         {
-            OptionChoose.Visibility = Visibility.Collapsed;
             AddForm.Visibility = Visibility.Visible;
             ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Collapsed;
         }
 
-        private void ModFormVisible(object sender, RoutedEventArgs e)
+        private void Cancel(object sender, RoutedEventArgs e)
         {
             AddForm.Visibility = Visibility.Collapsed;
-            OptionChoose.Visibility = Visibility.Collapsed;
-            ModForm.Visibility = Visibility.Visible;
-
-            DataTable temp = DBConnection.BasicId("[dbo].[ProcSelectIdProjekty]", int.Parse(selectedId));
-
-            MDatePicker1.SelectedDate = ((DateTime)temp.Rows[0][0]);
-            MOpisSource.Text = temp.Rows[0][1].ToString();
-            MKSource.SelectedValue = temp.Rows[0][2];
-            MZSource.SelectedValue = temp.Rows[0][3];
+            ModForm.Visibility = Visibility.Collapsed;
+            Filters.Visibility = Visibility.Visible;
         }
-
 
 
         private void AddNewRecord(object sender, RoutedEventArgs e)
@@ -133,18 +132,47 @@ namespace SQLProjektV2.Views
 
         private void DeleteRecord(object sender, RoutedEventArgs e)
         {
-            string temp = $"DELETE FROM [dbo].[Projekty] WHERE Id = {selectedId}";
-            try
+            MessageBoxResult result = MessageBox.Show("Czy na pewno chcesz usunąć wpis?", "", MessageBoxButton.YesNo);
+
+            if (result == MessageBoxResult.Yes)
             {
-                DBConnection.SQLCommand(temp);
-                MessageBox.Show("Usunięto projekt");
-                DataContext = new ProjektyViewModel();
-                OptionChoose.Visibility = Visibility.Collapsed;
+                string temp = $"DELETE FROM [dbo].[Projekty] WHERE Id = {selectedId}";
+                try
+                {
+                    DBConnection.SQLCommand(temp);
+                    MessageBox.Show("Usunięto projekt");
+                    DataContext = new ProjektyViewModel();
+                    ModForm.Visibility = Visibility.Collapsed;
+                    Filters.Visibility = Visibility.Visible;
+                }
+                catch (Exception)
+                {
+                    MessageBox.Show("Nie usunięto projektu, ponieważ jest związany z innymi tablicami");
+                }
             }
-            catch (Exception)
-            {
-                MessageBox.Show("Nie usunięto projektu, ponieważ jest związany z innymi tablicami");
-            }
+        }
+
+
+        private void AddFilter(object sender, RoutedEventArgs e)
+        {
+            StackPanel temp = Globals.StackPanel;
+
+            ComboBox tempCB = Globals.ComboBoxColumnChoose;
+            tempCB.ItemsSource = ((ProjektyViewModel)this.DataContext).FilterInfo;
+            temp.Children.Add(tempCB);
+
+            temp.Children.Add(Globals.ComboBox);
+
+            temp.Children.Add(Globals.Button);
+
+            (temp.Children[0] as ComboBox).SelectedIndex = 0;
+
+            FiltersList.Children.Add(temp);
+        }
+
+        private void UseFilters(object sender, RoutedEventArgs e)
+        {
+            (DataContext as ProjektyViewModel).MainTable.RowFilter = Globals.GetFilter(FiltersList);
         }
     }
 }
